+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 * 0.3.12 main cheat 1st release
 * 0.3.13 removed pollution from schools
 * 0.3.14 New github and Bepex6 integration
+* 0.3.15 Added:
+	* CoalPowerplant
+	* Nuclear Powerplant
+	* Windmill
+	* Battery
+	* Solar powerstation
+	* Geothermal